refactor: extract per-module metrics into analyzer/metrics.rs (#190)
Moves ModuleMetrics, ExternalDepMetric, and the hotspot/blast-radius
computation (now compute_hotspots) out of analyzer/mod.rs.
